Port to CMake

+++ cl_dll/hudgl.cpp
@@ -70,13 +70,14 @@ void HudGL::rectangle(const Vector2D& corner_a, const Vector2D& corner_b) const 
 std::vector<Vector2D> HudGL::compute_circle(float radius) {
 	// Maximum allowed distance between the circle and the rendered line segment.
 	constexpr float MAX_ERROR = 0.1f;
-	const unsigned segment_count = std::ceil(M_PI / std::acos((radius - MAX_ERROR) / radius));
+	const unsigned segment_count =
+		static_cast<unsigned>(std::ceil(M_PI / std::acos((radius - MAX_ERROR) / radius)));
 
 	std::vector<Vector2D> points;
 	points.reserve(segment_count);
 
 	for (unsigned i = 0; i < segment_count; ++i) {
-		float angle = M_PI * 2 * i / segment_count;
+		float angle = static_cast<float>(M_PI * 2 * i / segment_count);
 		points.emplace_back(radius * std::cos(angle), radius * std::sin(angle));
 	}
